+++ vantage-api-client/src/graphql/vista/source.rs
@@ -128,6 +128,12 @@ impl TableShell for GraphqlApiTableShell {
         factory.from_table(target)
     }
 
+    fn get_ref_target(&self, relation: &str) -> Result<Vista> {
+        let target = self.table.get_ref_target::<EmptyEntity>(relation)?;
+        let factory = GraphqlApiVistaFactory::new(self.table.data_source().clone());
+        factory.from_table(target)
+    }
+
     fn get_ref_kinds(&self) -> Vec<(String, ReferenceKind)> {
         self.table.ref_kinds()
     }